Extract functionality of basic SNN into own file
