+++ snn_hfo_ieeg/stages/snn/signal_enhancer.py
@@ -1,5 +1,6 @@
 import numpy as np
 from snn_hfo_ieeg.stages.snn.creation import create_synapses
+from snn_hfo_ieeg.stages.snn.basic_network import create_input_layer, create_input_to_hidden_synapses
 from snn_hfo_ieeg.user_facing_data import MeasurementMode
 
 
@@ -16,3 +17,19 @@ def create_signal_enhancer_to_output_synapses(signal_enhancer_layer, output_laye
 
 def get_signal_enhancer_input_bandwidth(filtered_spikes):
     return [filtered_spikes.ripple, filtered_spikes.above_fast_ripple]
+
+
+def add_input_to_signal_enhancer_to_network(cache, filtered_spikes):
+    signal_enhancer_filtered_bandwidths = get_signal_enhancer_input_bandwidth(
+        filtered_spikes)
+    signal_enhancer_input_layer = create_input_layer(
+        signal_enhancer_filtered_bandwidths, cache.neuron_counts.input)
+
+    signal_enhancer_input_to_hidden_synapses = create_input_to_hidden_synapses(
+        signal_enhancer_input_layer,
+        cache.signal_enhancer_hidden_layer,
+        cache.model_paths,
+        cache.neuron_counts)
+    cache.network.add(signal_enhancer_input_layer)
+    cache.network.add(signal_enhancer_input_to_hidden_synapses)
+    return signal_enhancer_input_layer, signal_enhancer_input_to_hidden_synapses
